+++ classes/extension/configuration.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace mageekguy\atoum\extension;
+
+
+interface configuration
+{
+	public function serialize();
+
+	public static function unserialize(array $configuration);
+}